+++ R/top.R
@@ -5,15 +5,19 @@
 #' @param n Number of top entries to show
 #' @param output Output format: vector or data.frame
 #' @param round Optional rounding
+#' @param na.rm Logical. Remove NA before calculating the statistics.
 #' @return Vector of top counts, named by the corresponding entries
 #' @export
 #' @author Leo Lahti \email{leo.lahti@@iki.fi}
 #' @references See citation("bibliographica")
 #' @examples \dontrun{p <- top(df, field, 10)}
 #' @keywords utilities
-top <- function (x, field = NULL, n = NULL, output = "vector", round = NULL) {
+top <- function (x, field = NULL, n = NULL, output = "vector", round = NULL, na.rm = FALSE) {
 
   if (is.vector(x)) {
+    if (na.rm) {
+      x <- x[!is.na(x)]
+    }
     s <- rev(sort(table(x)))
     N <- length(x)
   } else if (is.data.frame(x) || is.matrix(x)) {
@@ -21,8 +25,13 @@ top <- function (x, field = NULL, n = NULL, output = "vector", round = NULL) {
       #warning("Indicate the selected field for the top function.")
       return(NULL)
     }
-    N <- nrow(x)
-    s <- rev(sort(table(x[, field])))
+
+    x <- x[, field]
+    if (na.rm) {
+      x <- x[!is.na(x)]
+    }
+    N <- length(x)
+    s <- rev(sort(table(x)))
   }
 
   if (!is.null(n)) {