+++ NEWS.md
@@ -3,5 +3,6 @@ bibliographica 0.2.29 (2016-10-20)
 
 ### NEW FEATURES
 
+  * Added the na.rm option in the top function
   * Splitted the timeline function into plot_timeline (plot) and timeline (data).